populate edit form with tag info
