+++ scripts/journalEntryList.js
@@ -1,7 +1,8 @@
 import { journalEntryHTML } from "./journalEntry.js"
 import { getJournalEntries, useJournalEntries } from "./journalDataProvider.js"
-import { getTags, useTags } from "./tagProvider.js"
-import { getEntryTags, useEntryTags } from "./entryTagsProvider.js"
+import { getTags,  } from "./tagProvider.js"
+import { getEntryTags,  } from "./entryTagsProvider.js"
+import { findTagsByEntry } from "./tags.js"
 
 const contentTarget = document.querySelector(".past-entries")
 const eventHub = document.querySelector(".container")
@@ -28,21 +29,10 @@ export const listEntries = () => {
     .then(getEntryTags)
     .then( () => {
       const journalEntries = useJournalEntries()
-      const tags = useTags()
-      const entryTags = useEntryTags()
-
-      
 
       // convert the entry objects to html
       let journalEnriesAsHTML = journalEntries.map(entry => {
-        const relatedEntryTags = entryTags.filter( entryTag => {
-          return entryTag.entryId === entry.id
-        })
-
-        const relatedTags = relatedEntryTags.map( relatedEntryTag => {
-          return tags.find( tag => tag.id === relatedEntryTag.tagId )
-        })
-
+        const relatedTags = findTagsByEntry(entry.id)
         return journalEntryHTML(entry, relatedTags)
       }).join("")
